libutil: factor out delegate.c utility functions
Problem: some functions, like those for reading eventlog entries and
for printing job IDs, may be useful in many different contexts.

Factor them out into src/common.
