+++ src/common/Makefile.am
@@ -1,0 +1,2 @@
+SUBDIRS = \
+	libutil
