+++ src/common/libutil/eventlog.c
@@ -1,0 +1,116 @@
+/************************************************************\
+ * Copyright 2021 Lawrence Livermore National Security, LLC
+ * (c.f. AUTHORS, NOTICE.LLNS, COPYING)
+ *
+ * This file is part of the Flux resource manager framework.
+ * For details, see https://github.com/flux-framework.
+ *
+ * SPDX-License-Identifier: LGPL-3.0
+\************************************************************/
+
+#if HAVE_CONFIG_H
+#include "config.h"
+#endif
+
+#include <errno.h>
+#include <fcntl.h>
+#include <stdint.h>
+#include <stdlib.h>
+#include <unistd.h>
+#include <string.h>
+#include <jansson.h>
+#include <flux/core.h>
+
+#include "eventlog.h"
+
+bool eventlog_entry_validate (json_t *entry)
+{
+    json_t *name;
+    json_t *timestamp;
+    json_t *context;
+
+    if (!json_is_object (entry) || !(name = json_object_get (entry, "name"))
+        || !json_is_string (name) || !(timestamp = json_object_get (entry, "timestamp"))
+        || !json_is_number (timestamp))
+        return false;
+
+    if ((context = json_object_get (entry, "context"))) {
+        if (!json_is_object (context))
+            return false;
+    }
+
+    return true;
+}
+
+json_t *eventlog_entry_decode (const char *entry)
+{
+    int len;
+    char *ptr;
+    json_t *o;
+
+    if (!entry)
+        goto einval;
+
+    if (!(len = strlen (entry)))
+        goto einval;
+
+    if (entry[len - 1] != '\n')
+        goto einval;
+
+    if (entry[len - 1] != '\n')
+        goto einval;
+
+    ptr = strchr (entry, '\n');
+    if (ptr != &entry[len - 1])
+        goto einval;
+
+    if (!(o = json_loads (entry, JSON_ALLOW_NUL, NULL)))
+        goto einval;
+
+    if (!eventlog_entry_validate (o)) {
+        json_decref (o);
+        goto einval;
+    }
+
+    return o;
+
+einval:
+    errno = EINVAL;
+    return NULL;
+}
+
+int eventlog_entry_parse (json_t *entry, double *timestamp, const char **name, json_t **context)
+{
+    double t;
+    const char *n;
+    json_t *c;
+
+    if (!entry) {
+        errno = EINVAL;
+        return -1;
+    }
+
+    if (json_unpack (entry, "{ s:F s:s }", "timestamp", &t, "name", &n) < 0) {
+        errno = EINVAL;
+        return -1;
+    }
+
+    if (!json_unpack (entry, "{ s:o }", "context", &c)) {
+        if (!json_is_object (c)) {
+            errno = EINVAL;
+            return -1;
+        }
+    } else
+        c = NULL;
+
+    if (timestamp)
+        (*timestamp) = t;
+    if (name)
+        (*name) = n;
+    if (context)
+        (*context) = c;
+
+    return 0;
+}
+
+// vi:ts=4 sw=4 expandtab
